Atualizando separação dos Widget, ok.
