+++ lib/main.dart
@@ -1,23 +1,4 @@
+import 'package:conversor_moedas/ui/android/pages/material_app.dart';
 import 'package:flutter/material.dart';
 
-void main(){
-  runApp(MaterialApp(
-    home: HomePage(),
-  ));
-}
-
-class HomePage extends StatefulWidget {
-  @override
-  _HomePageState createState() => _HomePageState();
-}
-
-class _HomePageState extends State<HomePage> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text("Bem vindo"),
-      ),
-    );
-  }
-}
+void main() => runApp(MyMaterialApp());
