+++ lib/ui/android/pages/home_page.dart
@@ -1,0 +1,20 @@
+
+import 'package:flutter/material.dart';
+
+class HomePage extends StatefulWidget {
+  @override
+  _HomePageState createState() => _HomePageState();
+}
+
+class _HomePageState extends State<HomePage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text("Conversor de Moedas"),
+
+        centerTitle: true,
+      ),
+    );
+  }
+}
